+++ Assets/lilToon/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.3.4] - 2022-07-xx
+### Added
+- Warning when using variants with GrabPass or geometry shader in project for cluster
+- ChilloutVR avatar build optimization
+
 ### Fixed
 - Fixed an issue that could cause script errors
 - Fixed color value after optimization